Refactor Windows protocol handling by moving implementation to a separate file
